@flip1995 2021-02-19
Timings on 40c server 

- first 100 lines of this list, aka everything from line 102 onward removed, aka `error-chain` is the last line
- Clippy and clippy_dev precompiled. All crates already downloaded. 
- **Without rayon**

```
cargo dev-lintcheck  576.96s user 53.13s system 161% cpu 6:30.50 total
```

Size of target dirs:

```bash
$ du -sh target
4.2G    target
$ du -sh target/lintcheck
1.6G    target/lintcheck
$ du -sh target/lintcheck/shared_target_dir
1.5G    target/lintcheck/shared_target_dir
```

---

- first 100 lines of this list, aka everything from line 102 onward removed, aka `error-chain` is the last line
- Clippy and clippy_dev precompiled. All crates already downloaded. 
- **With rayon**

```
cargo dev-lintcheck  1165.68s user 99.59s system 1837% cpu 1:08.87 total
```


Size of target dirs:

```bash
$ du -sh target
4.7G    target
$ du -sh target/lintcheck
2.2G    target/lintcheck
$ du -sh target/lintcheck/shared_target_dir
2.1G    target/lintcheck/shared_target_dir
```

---

Overall a speed up of **5.67x**. So it is definitely worth it if you have a machine with many cores, but not that relevant on smaller machines.

So I would add a flag `--parallel-checks N` and only check in parallel if you specify that. (Maybe if `N` isn't specified keep the current behavior of using half of the cores)