@matthiaskrgr 2021-02-19
Lol, so we we throw in 40x the performance for a ~5-6x speedup :laughing: ...
Then again, you only checked the first 100 crates, that's almost only twice as much as the core count.
I assume the parallel version would be more efficient by a larger margin if we had checked the full 400 crates?

I wonder if we could get slightly better performance by checking crates with large deptrees first or if the difference is negligible.