@flip1995 2021-02-19
Well if you used 4c/8t before its more like 10x the power. I also looked at `htop` a bit and when using the sequential version only 4-8 cores were in use at the same time. The parallel version used all 40 cores throughout.


>  wonder if we could get slightly better performance by checking crates with large deptrees first or if the difference is negligible.

I don't think that this will help much. Computing the dep trees and the figuring out what to compile where and in which order sounds definitely too complicated.