@matthiaskrgr 2021-02-19
Extremely hacky and ugly way to figure out the total number of deps in crate needs: 
````rust
format!("{:?}", cargo_metadata::MetadataCommand::new().manifest_path("./Cargo.toml").exec().unwrap().resolve.unwrap().nodes).matches("dependencies: ").count();
````