@flip1995 2021-02-20
@bors r+

Thanks!

Does this command compute only the direct deps or the complete tree?