@matthiaskrgr 2021-02-20
It counts the number of deps in the entire dep tree.